pyproject.toml
[tool.black]
line-length = 110
target-version = ["py38"]

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_local_folder = "apps,utils,cashback_api,tests"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry]
name = "cashback_api"
version = "0.0.0"
description = "REST API to cashback dashboard"
authors = ["Silvio Leite <silviolleite@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
belogging = "*"
dj-static = "*"
dj-database-url = "*"
django = "*"
django-cache-url = "*"
django-extensions = "*"
django-filter = "*"
djangorestframework = "*"
djangorestframework-jwt = "*"
django-cors-headers = "*"
python-decouple = "*"
drf-jsonmask = "^0.1.0"
djangorestframework-simplejwt = "^4.6.0"
requests = "^2.25.0"
factory-boy = "^3.1.0"
responses = "^0.12.1"
PyYAML = "^5.3.1"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-benchmark = "*"
pytest-cache = "*"
pytest-cov = "*"
pytest-django = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"